+++ src/main/java/com/elmakers/mine/bukkit/plugins/spells/builtin/TorchSpell.java
@@ -76,13 +76,14 @@ public class TorchSpell extends Spell
 		boolean isAir = face.getType() == Material.AIR;
 		boolean isAttachmentSlippery = target.getType() == Material.GLASS || target.getType() == Material.ICE;
 		boolean replaceAttachment = target.getType() == Material.SNOW || target.getType() == Material.NETHERRACK || target.getType() == Material.SOUL_SAND;
-		boolean isWater = target.getType() == Material.STATIONARY_WATER || target.getType() == Material.WATER;
+		boolean isWater = face.getType() == Material.STATIONARY_WATER || face.getType() == Material.WATER;
 		boolean isNether = target.getType() == Material.NETHERRACK || target.getType() == Material.SOUL_SAND;
 		Material targetMaterial = Material.TORCH;
 		
 		if (isWater || isAttachmentSlippery || isNether)
 		{
 			targetMaterial = Material.GLOWSTONE;
+			replaceAttachment = true;
 		}
 		
 		if